Boolean Embree: use double precision

+++ apps/boolean_embree/cgal_types.hh
@@ -13,7 +13,7 @@
 #include "vec3.hh"
 
 // typedef CGAL::Exact_predicates_inexact_constructions_kernel K;
-typedef CGAL::Simple_cartesian<float> K;
+typedef CGAL::Simple_cartesian<double> K;
 typedef K::Triangle_3 Triangle;
 typedef K::Segment_3 Segment;
 typedef K::Point_3 Point;
